Rename hashGen.js -> commands.js
